рефакторит код, добавляет хуки сортировки, фильтрации,валидации цены
